[Validator] Fix registering "is_valid()" for `#[Expression]`
