+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/validator.php
@@ -15,6 +15,7 @@ use Symfony\Bundle\FrameworkBundle\CacheWarmer\ValidatorCacheWarmer;
 use Symfony\Component\Cache\Adapter\PhpArrayAdapter;
 use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
 use Symfony\Component\Validator\Constraints\EmailValidator;
+use Symfony\Component\Validator\Constraints\ExpressionLanguageProvider;
 use Symfony\Component\Validator\Constraints\ExpressionValidator;
 use Symfony\Component\Validator\Constraints\NoSuspiciousCharactersValidator;
 use Symfony\Component\Validator\Constraints\NotCompromisedPasswordValidator;
@@ -82,10 +83,15 @@ return static function (ContainerConfigurator $container) {
 
         ->set('validator.expression_language', ExpressionLanguage::class)
             ->args([service('cache.validator_expression_language')->nullOnInvalid()])
+            ->call('registerProvider', [
+                service('validator.expression_language_provider')->ignoreOnInvalid(),
+            ])
 
         ->set('cache.validator_expression_language')
             ->parent('cache.system')
             ->tag('cache.pool')
+
+        ->set('validator.expression_language_provider', ExpressionLanguageProvider::class)
 
         ->set('validator.email', EmailValidator::class)
             ->args([
