+++ src/Symfony/Component/Validator/Constraints/ExpressionValidator.php
@@ -11,8 +11,6 @@
 
 namespace Symfony\Component\Validator\Constraints;
 
-use Symfony\Component\ExpressionLanguage\ExpressionFunction;
-use Symfony\Component\ExpressionLanguage\ExpressionFunctionProviderInterface;
 use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
@@ -22,15 +20,14 @@ use Symfony\Component\Validator\Exception\UnexpectedTypeException;
  * @author Fabien Potencier <fabien@symfony.com>
  * @author Bernhard Schussek <bschussek@symfony.com>
  */
-class ExpressionValidator extends ConstraintValidator implements ExpressionFunctionProviderInterface
+class ExpressionValidator extends ConstraintValidator
 {
     private ExpressionLanguage $expressionLanguage;
 
     public function __construct(ExpressionLanguage $expressionLanguage = null)
     {
         if ($expressionLanguage) {
-            $this->expressionLanguage = clone $expressionLanguage;
-            $this->expressionLanguage->registerProvider($this);
+            $this->expressionLanguage = $expressionLanguage;
         }
     }
 
@@ -56,25 +53,11 @@ class ExpressionValidator extends ConstraintValidator implements ExpressionFunct
         }
     }
 
-    public function getFunctions(): array
-    {
-        return [
-            new ExpressionFunction('is_valid', function (...$arguments) {
-                return sprintf(
-                    '0 === $context->getValidator()->inContext($context)->validate(%s)->getViolations()->count()',
-                    implode(', ', $arguments)
-                );
-            }, function (array $variables, ...$arguments): bool {
-                return 0 === $variables['context']->getValidator()->inContext($variables['context'])->validate(...$arguments)->getViolations()->count();
-            }),
-        ];
-    }
-
     private function getExpressionLanguage(): ExpressionLanguage
     {
         if (!isset($this->expressionLanguage)) {
             $this->expressionLanguage = new ExpressionLanguage();
-            $this->expressionLanguage->registerProvider($this);
+            $this->expressionLanguage->registerProvider(new ExpressionLanguageProvider());
         }
 
         return $this->expressionLanguage;
